Initial commit
There's a lot going on here. We're basing the repository on [this
template](https://github.com/ColinKennedy/nvim-best-practices-plugin-template)
as a starting point. Since this is a project to incorporate beads into
Neovim, you won't be surprised to see beads integrated into this
repository. Finally, I've started some initial design work, pulling
together and initial design document and some technical references.

At this stage we're ready to start putting together an initial
high-level development plan, after which the plan will be broken down
into issues (yup, in beads), so we can begin the implementation.