Add a generic execute subcommand
This takes a lot of inspiration from vim-fugitive. The 'execute'
subcommand passes the supplied command and full set of arguments to the
'bd' executable, but then provides some conditional handling of the
results so that, for example, if the command returns an issue list, we
show that list in a Telescope picker.

This is the first step to making 'search' work as you can now run 'Beads
execute search' and it'll do what you expect.

The next commit will add a 'Beads search' porcelain on that to provide a
simplified interface for more common, basic search forms.